Extract challenge routes
