+++ orm/Dockerfile
@@ -5,7 +5,7 @@ RUN go mod download
 COPY utils/ utils/
 COPY model/ model/
 COPY orm/ orm/
-RUN go build orm/main.go
+RUN go build orm/main.go orm/challenge.go
 
 FROM alpine
 WORKDIR /go/src/app
